Protecting against the Balance Overflow (u64) at the receiver side
The above assert addition checks for the overflow of at the receiver side.
As anyone can send any amount of token to anyone (given it's valid with balance checks). If we don't have a check on the overflow attack they possibly can send the required amount of token (u64 Max Value - User's Current Balance) to overflow the user's balance to 0.
The aforementioned assert condition checks the Balance of user before and after the token addition. In no way we can have reduced token than before, If we get that condition it points that an overflow attack is being carried out. The assert condition precisely captures that and does not allow the transfer.